tetragon: push execObj() into new pkg to avoid circ dep
We will need to use execObj() from the policyfilter pkg, but this will
cause a circular dependency. So lets just pull this out into a new pkg
that we can import.
